python_services/requirements.txt
fastapi
uvicorn
openai-whisper
spacy
nltk
transformers
torch
